Split comments-related endpoints into a dedicated `comments` module:
- Moved comment-related routes (e.g., retrieval, creation, reactions) from `videos.py` to a new `comments.py`.
- Updated route prefixes and paths for consistency (`/api/comments`).
- Refactored dependencies and services for comments into `CommentService`.
- Adjusted `routers.py` to include the new `comments` router.
- Enhanced route structure by simplifying and organizing video and file endpoints.
