+++ backend/app/routers.py
@@ -1,4 +1,5 @@
 from src.api.auth import router_auth
+from src.api.comments import router_comments
 from src.api.files import router_files
 from src.api.health import router_health
 from src.api.metrics import router_metrics
@@ -12,6 +13,7 @@ def include_routers(app):
     app.include_router(router_files)
     app.include_router(router_metrics)
     app.include_router(router_videos)
+    app.include_router(router_comments)
     app.include_router(rabbit_router)
     app.include_router(router_search)
     app.include_router(router_auth)
